% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovCorPer.R
\name{CovCorPer}
\alias{CovCorPer}
\title{Covariance or correlation matrix from the Per-ACF}
\usage{
CovCorPer(x, type = c("correlation", "covariance"))
}
\arguments{
\item{x}{a numeric matrix}

\item{type}{character string giving the type of acf to be computed. Allowed values are "correlation" (the default) or "covariance".}
}
\value{
a numeric matrix
}
\description{
Wrapper that computes the covariance or correlation matrix of \code{x} at lag 0 obtained from the Per-ACF.
}
\examples{
data.set <- cbind(fdeaths, mdeaths)
CovCorPer(data.set)
}
