% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCrossPeriodogram.R
\name{MCrossPeriodogram}
\alias{MCrossPeriodogram}
\title{Robust M-cross-periodogram}
\usage{
MCrossPeriodogram(series1, series2)
}
\arguments{
\item{series1}{univariate time series}

\item{series2}{univariate time series}
}
\value{
a numeric vector containing the estimates of the cross-spectral density
}
\description{
This function computes the Robust M-cross-periodogram using M-regression.
}
\references{
Fuller, Wayne A. Introduction to statistical time series.  John Wiley & Sons, 2009.
}
\author{
Higor Cotta, Valdério A. Reisen, Pascal Bondon and Céline Lévy-Leduc
}
