% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerioReg.R
\name{PerioReg}
\alias{PerioReg}
\title{Periodogram}
\usage{
PerioReg(series)
}
\arguments{
\item{series}{univariate time series}
}
\value{
a numeric vector containing the robust estimates of the spectral density
}
\description{
This function computes the univariate periodogram using harmonic regression.
}
\examples{
PerioReg(ldeaths)
}
\references{
Reisen, V. A. and Lévy-Leduc, C. and Taqqu, M. (2017) An M-estimator for the long-memory parameter. \emph{Journal of Statistical Planning and Inference},  187, 44-55.

Fuller, Wayne A. Introduction to statistical time series.  John Wiley & Sons, 2009.
}
\author{
Higor Cotta, Valdério A. Reisen, Pascal Bondon and Céline Lévy-Leduc.
}
