% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rob_acf.R
\name{plot.robacf}
\alias{plot.robacf}
\title{Plot Robust Autocovariance and Robust Autocorrelation Functions}
\usage{
\method{plot}{robacf}(
  x,
  type = "h",
  xlab = "Lag",
  ylab = NULL,
  ylim = NULL,
  main = NULL,
  max.mfrow = 6,
  ask = Npgs > 1 && dev.interactive(),
  mar = if (nser > 2) c(3, 2, 2, 0.8) else par("mar"),
  oma = if (nser > 2) c(1, 1.2, 1, 1) else par("oma"),
  mgp = if (nser > 2) c(1.5, 0.6, 0) else par("mgp"),
  xpd = par("xpd"),
  cex.main = if (nser > 2) 1 else par("cex.main"),
  verbose = getOption("verbose"),
  ...
)
}
\arguments{
\item{x}{an object of class "robacf".}

\item{type}{the type of plot to be drawn, default to histogram like vertical lines.}

\item{xlab}{the x label of the plot.}

\item{ylab}{the y label of the plot.}

\item{ylim}{numeric of length 2 giving the y limits for the plot.}

\item{main}{overall title for the plot.}

\item{max.mfrow}{positive integer; for multivariate x indicating how many rows and columns of plots should be put on one page, using par(mfrow = c(m,m))(see \code{\link[graphics]{par}}).}

\item{ask}{logical; if TRUE, the user is asked before a new page is started.}

\item{mar, oma, mgp, xpd, cex.main}{graphics parameters as in par(*), by default adjusted to use smaller than default margins for multivariate x only.}

\item{verbose}{logical. Should R report extra information on progress?}

\item{...}{graphics parameters to be passed to the plotting routines.}
}
\value{
None
}
\description{
Plot method for objects of class "robacf". Mostly of the code re-used from the standard acf class.
}
\section{Contributions}{
 plot.acf (stats) - R Core
}

