% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TheoreticalACF.R
\name{TheoreticalACF}
\alias{TheoreticalACF}
\title{Compute the theoretical ACOVF or ACF for a VAR(1) process}
\usage{
TheoreticalACF(sig, phi, lagmax, s1, s2, type = c("correlation",
  "covariance"))
}
\arguments{
\item{sig}{a numeric white noise covariance matrix}

\item{phi}{a numeric coefficient matrix for the autoregressive order}

\item{lagmax}{maximum lag at which to calculate the acovf or acf}

\item{s1}{an interger containing the entry of the series in the vector series. if s1 == s2 the acovf or acf is calculated.}

\item{s2}{an interger containing the entry of the series in the vector series. if s1 != s2 the cross-acovf or cross-acf is calculated.}

\item{type}{character string giving the type to be computed. Allowed values are "correlation" (the default) or "covariance".}
}
\value{
a numeric matrix
}
\description{
Compute the theoretical autocovariance or autocorrelation function for a VAR(1) process.
}
\examples{
phi <- matrix(c(0.7,0.5,0,0.7),2,2)
cov.mat <- matrix(c(1,0,0,1),2,2)
lagmax <- 9
TheoreticalACF(cov.mat,phi,lagmax,1,1,"correlation")
TheoreticalACF(cov.mat,phi,lagmax,1,2,"correlation")
TheoreticalACF(cov.mat,phi,lagmax,2,1,"correlation")
TheoreticalACF(cov.mat,phi,lagmax,2,2,"correlation")
}
\references{
Multivariate Time Series Analysis: with R and Financial Applications by Ruey S. Tsay, Wiley, 2014.
}
\author{
Higor Cotta. Adapted from Tsay's book.
}
